backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
alembic>=1.13.0
geoalchemy2>=0.14.0
# Auth
argon2-cffi>=23.1.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
# NSPD Client
httpx>=0.25.0
pyproj>=3.6.0
shapely>=2.0.0
